requirements.txt
streamlit>=1.35.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0

# Testing
pytest>=8.0.0
